requirements.txt
PyQt5>=5.15.0
opencv-python>=4.8.0
pyzbar>=0.1.9
numpy>=1.24.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
